feat: move pan camera to tracking pin functionality from MapControls to PoiCard
