+++ app/_components/MapControls.tsx
@@ -3,10 +3,9 @@ import { NavigationControl, GeolocateControl, useMap } from "react-map-gl/maplib
 import {ImportantPinContext} from "./useContext/ImportantPinContext";
 import { Pin } from "../_utils/global";
 
-
 const MapControls = (): React.JSX.Element => {
-  const trackingPinContext = useContext(ImportantPinContext);
-  const {current: map} = useMap();
+  // const trackingPinContext = useContext(ImportantPinContext);
+  // const {current: map} = useMap();
 
   const geolocationRef = useRef<maplibregl.GeolocateControl|null>(null);
 
@@ -17,26 +16,26 @@ const MapControls = (): React.JSX.Element => {
     geolocationRef.current.trigger();
   }, [geolocationRef.current])
 
-  useEffect(() => {
-    if (!trackingPinContext) return;
-    if (!trackingPinContext.trackingPin) return;
-    console.table(trackingPinContext.trackingPin);
-      handlePanMapToTrackingPin(trackingPinContext.trackingPin);
-  },[trackingPinContext?.trackingPin])
+  // useEffect(() => {
+  //   if (!trackingPinContext) return;
+  //   if (!trackingPinContext.trackingPin) return;
+  //   console.table(trackingPinContext.trackingPin);
+  //     handlePanMapToTrackingPin(trackingPinContext.trackingPin);
+  // },[trackingPinContext?.trackingPin])
 
-  const handlePanMapToTrackingPin = (pin: Pin) => {
-    try {
-      if (!map) throw "Can't find map";
-      map.flyTo({
-        center: [pin.search_longitude, pin.search_latitude],
-        duration: 1000,
-        minZoom: 24,
-        zoom: 17
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  }
+  // const handlePanMapToTrackingPin = (pin: Pin) => {
+  //   try {
+  //     if (!map) throw "Can't find map";
+  //     map.flyTo({
+  //       center: [pin.search_longitude, pin.search_latitude],
+  //       duration: 1000,
+  //       minZoom: 24,
+  //       zoom: 17
+  //     });
+  //   } catch (error) {
+  //     console.error(error);
+  //   }
+  // }
 
   const handleGeolocateError = (error: GeolocationPositionError) => {
     console.log("geolocate error detected");
